Sighted now works

+++ src/lifish/collision_layers.hpp
@@ -6,7 +6,7 @@ namespace Game {
 
 namespace Layers {
 
-enum Layer : unsigned short {
+enum Layer : unsigned {
 	DEFAULT,
 	PLAYERS,
 	ENEMIES,
